Contains code for backend and front end for the software.
organizer.py - Organizes the data
pathandrow.py - Searches the data in case the attribute for search query
are path, row and date.
latandlon.py - Searches the data in case the attribute for search query
are latitude, longitude and date.
searchtypeone.html - HTML front end for pathandrow.py
searchtypetwo.html - HTML front end for latandlon.py
typeone.php - Stores the data entered in the typeone form in a text file
to be read by pathandrow.py
typetwo.php - Stores the data entered in the typetwo form in a text file
to be read by latandlon.py